Moved EnemyEnemy collision handling into enemycollision class
